Build js/shared dir as a separate library (update webpack confing and package.json)
[#184027461]
