+++ tsconfig.json
@@ -7,6 +7,7 @@
     "esModuleInterop": true,
     "target": "es2018",
     "allowJs": true,
+    "declaration": true,
     "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
     "experimentalDecorators": true,
